refactor: ScheduleService를 분리
- ScheduleService가 너무 길어지는 문제 때문에 트레이너와 트레이니의 작업으로 분류하고, 그 중에서 별도 작업을 분리함
- 조회 쿼리는 추후에 트레이너와 트레이니의 조회 쿼리를 분리하면서 각 서비스에 맞게 들어갈 예정
