+++ src/test/java/com/project/trainingdiary/service/ScheduleServiceTest.java
@@ -1,41 +1,15 @@
 package com.project.trainingdiary.service;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.project.trainingdiary.dto.request.AcceptScheduleRequestDto;
-import com.project.trainingdiary.dto.request.ApplyScheduleRequestDto;
-import com.project.trainingdiary.dto.request.CancelScheduleByTraineeRequestDto;
-import com.project.trainingdiary.dto.request.CancelScheduleByTrainerRequestDto;
-import com.project.trainingdiary.dto.request.OpenScheduleRequestDto;
-import com.project.trainingdiary.dto.request.RegisterScheduleRequestDto;
-import com.project.trainingdiary.dto.request.RejectScheduleRequestDto;
-import com.project.trainingdiary.dto.response.CancelScheduleByTraineeResponseDto;
-import com.project.trainingdiary.dto.response.CancelScheduleByTrainerResponseDto;
-import com.project.trainingdiary.dto.response.RegisterScheduleResponseDto;
 import com.project.trainingdiary.dto.response.ScheduleResponseDto;
-import com.project.trainingdiary.entity.PtContractEntity;
-import com.project.trainingdiary.entity.ScheduleEntity;
 import com.project.trainingdiary.entity.TraineeEntity;
 import com.project.trainingdiary.entity.TrainerEntity;
-import com.project.trainingdiary.exception.impl.PtContractNotEnoughSession;
-import com.project.trainingdiary.exception.impl.PtContractNotExistException;
-import com.project.trainingdiary.exception.impl.ScheduleAlreadyExistException;
-import com.project.trainingdiary.exception.impl.ScheduleNotFoundException;
 import com.project.trainingdiary.exception.impl.ScheduleRangeTooLong;
-import com.project.trainingdiary.exception.impl.ScheduleStartIsPast;
-import com.project.trainingdiary.exception.impl.ScheduleStartTooSoon;
-import com.project.trainingdiary.exception.impl.ScheduleStartWithin1Day;
-import com.project.trainingdiary.exception.impl.ScheduleStatusNotOpenException;
-import com.project.trainingdiary.exception.impl.ScheduleStatusNotReserveApplied;
-import com.project.trainingdiary.exception.impl.ScheduleStatusNotReserveAppliedOrReserved;
-import com.project.trainingdiary.exception.impl.UsedSessionExceededTotalSession;
 import com.project.trainingdiary.model.ScheduleDateTimes;
 import com.project.trainingdiary.model.ScheduleResponseDetail;
 import com.project.trainingdiary.model.ScheduleStatus;
@@ -50,16 +24,13 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -228,102 +199,6 @@ class ScheduleServiceTest {
   }
 
   @Test
-  @DisplayName("일정 열기 - 성공(6개의 일정 열기)")
-  void openSchedule() {
-    //given
-    setupTrainerAuth();
-    List<ScheduleDateTimes> dateTimes = List.of(
-        ScheduleDateTimes.builder()
-            .startDate(LocalDate.of(2024, 1, 1))
-            .startTimes(List.of(
-                LocalTime.of(10, 0),
-                LocalTime.of(11, 0),
-                LocalTime.of(12, 0)
-            ))
-            .build(),
-        ScheduleDateTimes.builder()
-            .startDate(LocalDate.of(2024, 2, 28))
-            .startTimes(List.of(
-                LocalTime.of(20, 0),
-                LocalTime.of(21, 0),
-                LocalTime.of(22, 0)
-            ))
-            .build()
-    );
-    OpenScheduleRequestDto dto = OpenScheduleRequestDto.builder()
-        .dateTimes(dateTimes)
-        .build();
-
-    when(scheduleService.getScheduleList(
-        LocalDate.of(2024, 1, 1),
-        LocalDate.of(2024, 3, 1)
-    ))
-        .thenReturn(responseData);
-
-    //when
-    scheduleService.createSchedule(dto);
-    List<ScheduleResponseDto> schedules = scheduleService.getScheduleList(
-        LocalDate.of(2024, 1, 1),
-        LocalDate.of(2024, 3, 1)
-    );
-
-    //then
-    assertEquals(2, schedules.size());
-    assertEquals(3,
-        schedules.stream()
-            .filter(s -> s.getStartDate().equals(LocalDate.of(2024, 2, 28)))
-            .map(ScheduleResponseDto::getDetails)
-            .flatMap(List::stream)
-            .toList()
-            .size()
-    );
-  }
-
-  @Test
-  @DisplayName("일정 열기 - 실패(이미 일정이 있는 경우)")
-  void openScheduleFail_AlreadyExistSchedule() {
-    //given
-    setupTrainerAuth();
-    List<ScheduleDateTimes> dateTimes = List.of(
-        ScheduleDateTimes.builder()
-            .startDate(LocalDate.of(2024, 1, 1))
-            .startTimes(List.of(
-                LocalTime.of(10, 0),
-                LocalTime.of(11, 0),
-                LocalTime.of(12, 0)
-            ))
-            .build(),
-        ScheduleDateTimes.builder()
-            .startDate(LocalDate.of(2024, 2, 28))
-            .startTimes(List.of(
-                LocalTime.of(20, 0), // <- 중복된 일정이 들어가 있음
-                LocalTime.of(21, 0),
-                LocalTime.of(22, 0)
-            ))
-            .build()
-    );
-
-    OpenScheduleRequestDto dto = OpenScheduleRequestDto.builder()
-        .dateTimes(dateTimes)
-        .build();
-
-    when(scheduleRepository.findScheduleDatesByDates(
-        eq(LocalDateTime.of(2024, 1, 1, 10, 0)),
-        eq(LocalDateTime.of(2024, 2, 28, 22, 0))
-    ))
-        .thenReturn(Set.of(
-            LocalDateTime.of(2024, 2, 28, 20, 0)
-        ));
-
-    //when
-    //then
-    assertThrows(
-        ScheduleAlreadyExistException.class,
-        () -> scheduleService.createSchedule(dto)
-    );
-  }
-
-  @Test
   @DisplayName("일정 목록 조회 - 성공")
   void getScheduleList() {
     when(scheduleRepository.getScheduleList(
@@ -348,915 +223,5 @@ class ScheduleServiceTest {
             LocalDate.of(2024, 9, 1)
         )
     );
-  }
-
-  @Test
-  @DisplayName("일정 닫기 - 성공")
-  void closeSchedule() {
-    //given
-    List<Long> scheduleIds = List.of(1L, 2L, 3L);
-    when(scheduleRepository.findAllById(scheduleIds))
-        .thenReturn(
-            List.of(
-                ScheduleEntity.builder().id(1L).scheduleStatus(ScheduleStatus.OPEN).build(),
-                ScheduleEntity.builder().id(2L).scheduleStatus(ScheduleStatus.OPEN).build(),
-                ScheduleEntity.builder().id(3L).scheduleStatus(ScheduleStatus.OPEN).build()
-            )
-        );
-
-    //when
-    scheduleService.closeSchedules(scheduleIds);
-    ArgumentCaptor<List<ScheduleEntity>> captor = ArgumentCaptor.forClass(List.class);
-
-    //then
-    verify(scheduleRepository).deleteAll(captor.capture());
-    assertEquals(3, captor.getValue().size());
-  }
-
-  @Test
-  @DisplayName("일정 닫기 - 실패(스케쥴 아이디 목록 중에 없는 스케쥴이 있음)")
-  void closeScheduleFail_ScheduleNotFound() {
-    //given
-    List<Long> scheduleIds = List.of(1L, 2L, 3L);
-
-    //when
-    when(scheduleRepository.findAllById(scheduleIds))
-        .thenReturn(
-            List.of(
-                ScheduleEntity.builder().id(1L).scheduleStatus(ScheduleStatus.OPEN).build(),
-                ScheduleEntity.builder().id(3L).scheduleStatus(ScheduleStatus.OPEN).build()
-            )
-        );
-
-    //then
-    assertThrows(
-        ScheduleNotFoundException.class,
-        () -> scheduleService.closeSchedules(scheduleIds)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 닫기 - 실패(스케쥴이 OPEN 상태가 아님)")
-  void closeScheduleFail_ScheduleNotOpen() {
-    //given
-    List<Long> scheduleIds = List.of(1L, 2L, 3L);
-
-    //when
-    when(scheduleRepository.findAllById(scheduleIds))
-        .thenReturn(
-            List.of(
-                ScheduleEntity.builder().id(1L).scheduleStatus(ScheduleStatus.OPEN).build(),
-                ScheduleEntity.builder().id(2L).scheduleStatus(ScheduleStatus.RESERVED).build(),
-                ScheduleEntity.builder().id(3L).scheduleStatus(ScheduleStatus.OPEN).build()
-            )
-        );
-
-    //then
-    assertThrows(
-        ScheduleStatusNotOpenException.class,
-        () -> scheduleService.closeSchedules(scheduleIds)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 예약 신청 - 성공")
-  void applySchedule() {
-    //given
-    setupTraineeAuth();
-    ApplyScheduleRequestDto dto = new ApplyScheduleRequestDto();
-    dto.setScheduleId(100L);
-    LocalDateTime currentTime = LocalDateTime.now();
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.OPEN)
-                .startAt(currentTime.plusHours(2).withMinute(0).withSecond(0).withNano(0))
-                .trainer(trainer)
-                .build()
-        ));
-
-    when(traineeRepository.findByEmail("trainee@example.com"))
-        .thenReturn(Optional.of(trainee));
-
-    when(ptContractRepository.findByTrainerIdAndTraineeId(1L, 10L))
-        .thenReturn(Optional.of(
-            PtContractEntity.builder()
-                .id(1000L)
-                .trainer(trainer)
-                .trainee(trainee)
-                .build()
-        ));
-
-    ArgumentCaptor<ScheduleEntity> captor = ArgumentCaptor.forClass(ScheduleEntity.class);
-    scheduleService.applySchedule(dto, currentTime);
-
-    //then
-    verify(scheduleRepository).save(captor.capture());
-    assertEquals(1000L, captor.getValue().getPtContract().getId());
-  }
-
-  @Test
-  @DisplayName("일정 예약 신청 - 실패(일정이 없는 경우)")
-  void applyScheduleFail_NoSchedule() {
-    //given
-    setupTraineeAuth();
-    ApplyScheduleRequestDto dto = new ApplyScheduleRequestDto();
-    dto.setScheduleId(100L);
-    LocalDateTime currentTime = LocalDateTime.now();
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.empty());
-
-    //then
-    assertThrows(
-        ScheduleNotFoundException.class,
-        () -> scheduleService.applySchedule(dto, currentTime)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 예약 신청 - 실패(둘이 연결된 계약이 없는 경우)")
-  void applyScheduleFail_NoPtContract() {
-    //given
-    setupTraineeAuth();
-    ApplyScheduleRequestDto dto = new ApplyScheduleRequestDto();
-    dto.setScheduleId(100L);
-    LocalDateTime currentTime = LocalDateTime.now();
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.OPEN)
-                .startAt(currentTime.plusHours(2).withMinute(0).withSecond(0).withNano(0))
-                .trainer(trainer)
-                .build()
-        ));
-
-    when(ptContractRepository.findByTrainerIdAndTraineeId(1L, 10L))
-        .thenReturn(Optional.empty());
-
-    //then
-    assertThrows(
-        PtContractNotExistException.class,
-        () -> scheduleService.applySchedule(dto, currentTime)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 예약 신청 - 실패(OPEN 일정이 아닌 경우)")
-  void applyScheduleFail_ScheduleNotOpen() {
-    //given
-    setupTraineeAuth();
-    ApplyScheduleRequestDto dto = new ApplyScheduleRequestDto();
-    dto.setScheduleId(100L);
-    LocalDateTime currentTime = LocalDateTime.now();
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.RESERVED)
-                .startAt(currentTime.plusHours(2).withMinute(0).withSecond(0).withNano(0))
-                .trainer(trainer)
-                .build()
-        ));
-
-    //then
-    assertThrows(
-        ScheduleStatusNotOpenException.class,
-        () -> scheduleService.applySchedule(dto, currentTime)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 예약 신청 - 실패(과거의 일정인 경우)")
-  void applyScheduleFail_ScheduleIsPast() {
-    //given
-    setupTraineeAuth();
-    ApplyScheduleRequestDto dto = new ApplyScheduleRequestDto();
-    dto.setScheduleId(100L);
-    LocalDateTime currentTime = LocalDateTime.now();
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.OPEN)
-                .startAt(currentTime.minusHours(2).withMinute(0).withSecond(0).withNano(0))
-                .trainer(trainer)
-                .build()
-        ));
-
-    //then
-    assertThrows(
-        ScheduleStartIsPast.class,
-        () -> scheduleService.applySchedule(dto, currentTime)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 예약 신청 - 실패(1시간 이내 시작하는 일정인 경우)")
-  void applyScheduleFail_ScheduleTooSoon() {
-    //given
-    setupTraineeAuth();
-    ApplyScheduleRequestDto dto = new ApplyScheduleRequestDto();
-    dto.setScheduleId(100L);
-    LocalDateTime currentTime = LocalDateTime.now();
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.OPEN)
-                .startAt(currentTime.plusHours(1).withMinute(0).withSecond(0).withNano(0))
-                .trainer(trainer)
-                .build()
-        ));
-
-    //then
-    assertThrows(
-        ScheduleStartTooSoon.class,
-        () -> scheduleService.applySchedule(dto, currentTime)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 수락 - 성공")
-  void acceptSchedule() {
-    //given
-    setupTrainerAuth();
-    AcceptScheduleRequestDto dto = new AcceptScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.RESERVE_APPLIED)
-                .ptContract(
-                    PtContractEntity.builder()
-                        .id(1000L)
-                        .totalSession(10)
-                        .usedSession(5)
-                        .build()
-                )
-                .build()
-        ));
-
-    //then
-    scheduleService.acceptSchedule(dto);
-  }
-
-  @Test
-  @DisplayName("일정 수락 - 실패(없는 일정)")
-  void acceptScheduleFail_NoSchedule() {
-    //given
-    setupTrainerAuth();
-    AcceptScheduleRequestDto dto = new AcceptScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.empty());
-
-    //then
-    assertThrows(
-        ScheduleNotFoundException.class,
-        () -> scheduleService.acceptSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 수락 - 실패(아무도 신청하지 않은 예약을 수락하려 함)")
-  void acceptScheduleFail_NoReserveApplied() {
-    //given
-    setupTrainerAuth();
-    AcceptScheduleRequestDto dto = new AcceptScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.OPEN)
-                .ptContract(
-                    PtContractEntity.builder()
-                        .id(1000L)
-                        .totalSession(10)
-                        .usedSession(10) // 모두 사용
-                        .build()
-                )
-                .build()
-        ));
-
-    //then
-    assertThrows(
-        ScheduleStatusNotReserveApplied.class,
-        () -> scheduleService.acceptSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 수락 - 실패(연결된 PT 계약이 없음)")
-  void acceptScheduleFail_NoPtContract() {
-    //given
-    setupTrainerAuth();
-    AcceptScheduleRequestDto dto = new AcceptScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.RESERVE_APPLIED)
-                .ptContract(null)
-                .build()
-        ));
-
-    //then
-    assertThrows(
-        PtContractNotExistException.class,
-        () -> scheduleService.acceptSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 수락 - 실패(모든 세션 횟수를 사용함)")
-  void acceptScheduleFail_UsedAllSession() {
-    //given
-    setupTrainerAuth();
-    AcceptScheduleRequestDto dto = new AcceptScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.RESERVE_APPLIED)
-                .ptContract(
-                    PtContractEntity.builder()
-                        .id(1000L)
-                        .totalSession(10)
-                        .usedSession(10) // 모두 사용
-                        .build()
-                )
-                .build()
-        ));
-
-    //then
-    assertThrows(
-        UsedSessionExceededTotalSession.class,
-        () -> scheduleService.acceptSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 거절 - 성공")
-  void rejectSchedule() {
-    //given
-    setupTrainerAuth();
-    RejectScheduleRequestDto dto = new RejectScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.RESERVE_APPLIED)
-                .ptContract(
-                    PtContractEntity.builder()
-                        .id(1000L)
-                        .totalSession(10)
-                        .usedSession(5)
-                        .build()
-                )
-                .build()
-        ));
-
-    ArgumentCaptor<PtContractEntity> captor = ArgumentCaptor.forClass(PtContractEntity.class);
-    scheduleService.rejectSchedule(dto);
-
-    //then
-    verify(ptContractRepository).save(captor.capture());
-    assertEquals(4, captor.getValue().getUsedSession());
-  }
-
-  @Test
-  @DisplayName("일정 거절 - 실패(없는 일정)")
-  void rejectScheduleFail_NoSchedule() {
-    //given
-    setupTrainerAuth();
-    RejectScheduleRequestDto dto = new RejectScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.empty());
-
-    //then
-    assertThrows(
-        ScheduleNotFoundException.class,
-        () -> scheduleService.rejectSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 거절 - 실패(아무도 신청하지 않은 예약을 거절하려 함)")
-  void rejectScheduleFail_NoReserveApplied() {
-    //given
-    setupTrainerAuth();
-    RejectScheduleRequestDto dto = new RejectScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.OPEN)
-                .ptContract(
-                    PtContractEntity.builder()
-                        .id(1000L)
-                        .totalSession(10)
-                        .usedSession(5)
-                        .build()
-                )
-                .build()
-        ));
-
-    //then
-    assertThrows(
-        ScheduleStatusNotReserveApplied.class,
-        () -> scheduleService.rejectSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 거절 - 실패(연결된 PT 계약이 없음)")
-  void rejectScheduleFail_NoPtContract() {
-    //given
-    setupTrainerAuth();
-    RejectScheduleRequestDto dto = new RejectScheduleRequestDto();
-    dto.setScheduleId(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.RESERVE_APPLIED)
-                .ptContract(null)
-                .build()
-        ));
-
-    //then
-    assertThrows(
-        PtContractNotExistException.class,
-        () -> scheduleService.rejectSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 등록 - 성공(OPEN된 일정이 없는 경우)")
-  void registerScheduleSuccess_NoOpen() {
-    //given
-    setupTrainerAuth();
-    RegisterScheduleRequestDto dto = new RegisterScheduleRequestDto(trainee.getId(),
-        dateTimesForRegister);
-
-    //when
-    when(ptContractRepository.findByTrainerIdAndTraineeId(1L, 10L))
-        .thenReturn(Optional.of(
-            PtContractEntity.builder()
-                .id(1000L)
-                .trainer(trainer)
-                .trainee(trainee)
-                .totalSession(10)
-                .usedSession(0)
-                .isTerminated(false)
-                .build()
-        ));
-
-    when(scheduleRepository.findScheduleDatesByDates(
-        LocalDateTime.of(2024, 1, 2, 20, 0),
-        LocalDateTime.of(2024, 1, 8, 22, 0)
-    ))
-        .thenReturn(new HashSet<>());
-
-    ArgumentCaptor<List<ScheduleEntity>> captorSchedule = ArgumentCaptor.forClass(List.class);
-    ArgumentCaptor<PtContractEntity> captorPtContract = ArgumentCaptor.forClass(
-        PtContractEntity.class);
-
-    //then
-    RegisterScheduleResponseDto response = scheduleService.registerSchedule(dto);
-
-    verify(scheduleRepository, times(2)).saveAll(captorSchedule.capture());
-    verify(ptContractRepository).save(captorPtContract.capture());
-
-    assertEquals(3, captorSchedule.getAllValues().get(0).size());
-    assertEquals(0, captorSchedule.getAllValues().get(1).size());
-    assertEquals(3, captorPtContract.getValue().getUsedSession());
-    assertEquals(7, response.getRemainSession());
-  }
-
-  @Test
-  @DisplayName("일정 등록 - 성공(이미 OPEN된 일정이 있는 경우라도 성공)")
-  void registerScheduleSuccess_Open() {
-    //given
-    setupTrainerAuth();
-    RegisterScheduleRequestDto dto = new RegisterScheduleRequestDto(trainee.getId(),
-        dateTimesForRegister);
-
-    //when
-    when(ptContractRepository.findByTrainerIdAndTraineeId(1L, 10L))
-        .thenReturn(Optional.of(
-            PtContractEntity.builder()
-                .id(1000L)
-                .trainer(trainer)
-                .trainee(trainee)
-                .totalSession(10)
-                .usedSession(0)
-                .isTerminated(false)
-                .build()
-        ));
-
-    when(scheduleRepository.findScheduleDatesByDates(
-        LocalDateTime.of(2024, 1, 2, 20, 0),
-        LocalDateTime.of(2024, 1, 8, 22, 0)
-    ))
-        .thenReturn(new HashSet<>(
-            List.of(LocalDateTime.of(2024, 1, 2, 20, 0))
-        ));
-
-    when(scheduleRepository.findByDates(
-        LocalDateTime.of(2024, 1, 2, 20, 0),
-        LocalDateTime.of(2024, 1, 2, 20, 0)
-    ))
-        .thenReturn(
-            List.of(
-                ScheduleEntity.builder()
-                    .id(100L)
-                    .startAt(LocalDateTime.of(2024, 1, 2, 20, 0))
-                    .scheduleStatus(ScheduleStatus.OPEN)
-                    .trainer(trainer)
-                    .build()
-            )
-        );
-
-    ArgumentCaptor<List<ScheduleEntity>> captorSchedule = ArgumentCaptor.forClass(List.class);
-    ArgumentCaptor<PtContractEntity> captorPtContract = ArgumentCaptor.forClass(
-        PtContractEntity.class);
-
-    //then
-    RegisterScheduleResponseDto response = scheduleService.registerSchedule(dto);
-
-    verify(scheduleRepository, times(2)).saveAll(captorSchedule.capture());
-    verify(ptContractRepository).save(captorPtContract.capture());
-
-    assertEquals(2, captorSchedule.getAllValues().get(0).size());
-    assertEquals(1, captorSchedule.getAllValues().get(1).size());
-    assertEquals(3, captorPtContract.getValue().getUsedSession());
-    assertEquals(7, response.getRemainSession());
-  }
-
-  @Test
-  @DisplayName("일정 등록 - 실패(OPEN 상태가 아닌 일정이 이미 있는 경우)")
-  void registerScheduleFail_ScheduleExistButNotOpen() {
-    //given
-    setupTrainerAuth();
-    RegisterScheduleRequestDto dto = new RegisterScheduleRequestDto(trainee.getId(),
-        dateTimesForRegister);
-
-    //when
-    when(ptContractRepository.findByTrainerIdAndTraineeId(1L, 10L))
-        .thenReturn(Optional.of(
-            PtContractEntity.builder()
-                .id(1000L)
-                .trainer(trainer)
-                .trainee(trainee)
-                .totalSession(10)
-                .usedSession(0)
-                .isTerminated(false)
-                .build()
-        ));
-
-    when(scheduleRepository.findScheduleDatesByDates(
-        LocalDateTime.of(2024, 1, 2, 20, 0),
-        LocalDateTime.of(2024, 1, 8, 22, 0)
-    ))
-        .thenReturn(new HashSet<>(
-            List.of(LocalDateTime.of(2024, 1, 2, 20, 0))
-        ));
-
-    when(scheduleRepository.findByDates(
-        LocalDateTime.of(2024, 1, 2, 20, 0),
-        LocalDateTime.of(2024, 1, 2, 20, 0)
-    ))
-        .thenReturn(
-            List.of(
-                ScheduleEntity.builder()
-                    .id(100L)
-                    .startAt(LocalDateTime.of(2024, 1, 2, 20, 0))
-                    .scheduleStatus(ScheduleStatus.RESERVED)
-                    .trainer(trainer)
-                    .build()
-            )
-        );
-
-    //then
-    assertThrows(
-        ScheduleStatusNotOpenException.class,
-        () -> scheduleService.registerSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 등록 - 실패(PT 계약이 없는 경우)")
-  void registerScheduleFail_NoPtContract() {
-    //given
-    setupTrainerAuth();
-    RegisterScheduleRequestDto dto = new RegisterScheduleRequestDto(trainee.getId(),
-        dateTimesForRegister);
-
-    //when
-    when(ptContractRepository.findByTrainerIdAndTraineeId(1L, 10L))
-        .thenReturn(Optional.empty());
-
-    //then
-    assertThrows(
-        PtContractNotExistException.class,
-        () -> scheduleService.registerSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("일정 등록 - 실패(PT 횟수가 부족한 경우)")
-  void registerScheduleFail_SessionNotEnough() {
-    //given
-    setupTrainerAuth();
-    RegisterScheduleRequestDto dto = new RegisterScheduleRequestDto(trainee.getId(),
-        dateTimesForRegister);
-
-    //when
-    when(ptContractRepository.findByTrainerIdAndTraineeId(1L, 10L))
-        .thenReturn(Optional.of(
-            PtContractEntity.builder()
-                .id(1000L)
-                .trainer(trainer)
-                .trainee(trainee)
-                .totalSession(10)
-                .usedSession(8)
-                .isTerminated(false)
-                .build()
-        ));
-
-    when(scheduleRepository.findScheduleDatesByDates(
-        LocalDateTime.of(2024, 1, 2, 20, 0),
-        LocalDateTime.of(2024, 1, 8, 22, 0)
-    ))
-        .thenReturn(new HashSet<>(
-            List.of(LocalDateTime.of(2024, 1, 2, 20, 0))
-        ));
-
-    //then
-    assertThrows(
-        PtContractNotEnoughSession.class,
-        () -> scheduleService.registerSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("트레이너의 일정 취소 - 성공")
-  void cancelScheduleByTrainer() {
-    //given
-    setupTrainerAuth();
-    CancelScheduleByTrainerRequestDto dto = new CancelScheduleByTrainerRequestDto(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.RESERVED)
-                .trainer(trainer)
-                .ptContract(
-                    PtContractEntity.builder()
-                        .id(1000L)
-                        .totalSession(10)
-                        .usedSession(5)
-                        .build()
-                )
-                .build()
-        ));
-
-    CancelScheduleByTrainerResponseDto response = scheduleService.cancelSchedule(
-        dto);
-    ArgumentCaptor<PtContractEntity> ptContractCaptor = ArgumentCaptor.forClass(
-        PtContractEntity.class);
-    ArgumentCaptor<ScheduleEntity> scheduleCaptor = ArgumentCaptor.forClass(ScheduleEntity.class);
-
-    //then
-    verify(ptContractRepository).save(ptContractCaptor.capture());
-    verify(scheduleRepository).save(scheduleCaptor.capture());
-
-    assertEquals(4, ptContractCaptor.getValue().getUsedSession());
-    assertEquals(ScheduleStatus.OPEN, scheduleCaptor.getValue().getScheduleStatus());
-    assertEquals(ScheduleStatus.OPEN, response.getScheduleStatus());
-  }
-
-  @Test
-  @DisplayName("트레이너의 일정 취소 - 실패(스케쥴이 없음)")
-  void cancelScheduleByTrainerFail_NoSchedule() {
-    //given
-    setupTrainerAuth();
-    CancelScheduleByTrainerRequestDto dto = new CancelScheduleByTrainerRequestDto(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.empty());
-
-    //then
-    assertThrows(
-        ScheduleNotFoundException.class,
-        () -> scheduleService.cancelSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("트레이너의 일정 취소 - 실패(일정 상태가 OPEN이라 취소할 수 없음)")
-  void cancelScheduleByTrainerFail_OpenStatus() {
-    //given
-    setupTrainerAuth();
-    CancelScheduleByTrainerRequestDto dto = new CancelScheduleByTrainerRequestDto(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .scheduleStatus(ScheduleStatus.OPEN)
-                .trainer(trainer)
-                .ptContract(
-                    PtContractEntity.builder()
-                        .id(1000L)
-                        .totalSession(10)
-                        .usedSession(5)
-                        .build()
-                )
-                .build())
-        );
-
-    //then
-    assertThrows(
-        ScheduleStatusNotReserveAppliedOrReserved.class,
-        () -> scheduleService.cancelSchedule(dto)
-    );
-  }
-
-  @Test
-  @DisplayName("트레이니의 일정 취소 - 성공")
-  void cancelScheduleByTrainee() {
-    //given
-    setupTraineeAuth();
-    CancelScheduleByTraineeRequestDto dto = new CancelScheduleByTraineeRequestDto(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(
-            ScheduleEntity.builder()
-                .id(100L)
-                .startAt(LocalDateTime.of(2024, 7, 16, 9, 0, 0))
-                .scheduleStatus(ScheduleStatus.RESERVED)
-                .trainer(trainer)
-                .ptContract(
-                    PtContractEntity.builder()
-                        .id(1000L)
-                        .totalSession(10)
-                        .usedSession(5)
-                        .trainer(trainer)
-                        .trainee(trainee)
-                        .build()
-                )
-                .build()
-        ));
-
-    CancelScheduleByTraineeResponseDto response = scheduleService.cancelSchedule(
-        dto,
-        LocalDateTime.of(2024, 7, 15, 8, 0, 0)
-    );
-    ArgumentCaptor<PtContractEntity> ptContractCaptor = ArgumentCaptor.forClass(
-        PtContractEntity.class);
-    ArgumentCaptor<ScheduleEntity> scheduleCaptor = ArgumentCaptor.forClass(ScheduleEntity.class);
-
-    //then
-    verify(ptContractRepository).save(ptContractCaptor.capture());
-    verify(scheduleRepository).save(scheduleCaptor.capture());
-
-    assertEquals(4, ptContractCaptor.getValue().getUsedSession());
-    assertEquals(ScheduleStatus.OPEN, scheduleCaptor.getValue().getScheduleStatus());
-    assertEquals(ScheduleStatus.OPEN, response.getScheduleStatus());
-  }
-
-  @Test
-  @DisplayName("트레이니의 일정 취소 - 실패(스케쥴이 없음)")
-  void cancelScheduleByTraineeFail_NoSchedule() {
-    //given
-    setupTraineeAuth();
-    CancelScheduleByTraineeRequestDto dto = new CancelScheduleByTraineeRequestDto(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.empty());
-
-    //then
-    assertThrows(
-        ScheduleNotFoundException.class,
-        () -> scheduleService.cancelSchedule(
-            dto,
-            LocalDateTime.of(2024, 7, 15, 8, 0, 0)
-        ));
-  }
-
-  @Test
-  @DisplayName("트레이니의 일정 취소 - 실패(일정 상태가 OPEN이라 취소할 수 없음)")
-  void cancelScheduleByTraineeFail_OpenStatus() {
-    //given
-    setupTraineeAuth();
-    CancelScheduleByTraineeRequestDto dto = new CancelScheduleByTraineeRequestDto(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(ScheduleEntity.builder()
-            .id(100L)
-            .scheduleStatus(ScheduleStatus.OPEN)
-            .trainer(trainer)
-            .ptContract(
-                PtContractEntity.builder()
-                    .id(1000L)
-                    .trainer(trainer)
-                    .trainee(trainee)
-                    .totalSession(10)
-                    .usedSession(5)
-                    .build()
-            )
-            .build())
-        );
-
-    //then
-    assertThrows(
-        ScheduleStatusNotReserveAppliedOrReserved.class,
-        () -> scheduleService.cancelSchedule(
-            dto,
-            LocalDateTime.of(2024, 7, 15, 8, 0, 0)
-        ));
-  }
-
-  @Test
-  @DisplayName("트레이니의 일정 취소 - 실패(RESERVED 상태의 일정이 24시간 미만으로 남았음)")
-  void cancelScheduleByTraineeFail_ReservedWithin1Day() {
-    //given
-    setupTraineeAuth();
-    CancelScheduleByTraineeRequestDto dto = new CancelScheduleByTraineeRequestDto(100L);
-
-    //when
-    when(scheduleRepository.findById(100L))
-        .thenReturn(Optional.of(ScheduleEntity.builder()
-            .id(100L)
-            .startAt(LocalDateTime.of(2024, 7, 16, 7, 0, 0)) // 내일 아침 7시 예약
-            .scheduleStatus(ScheduleStatus.RESERVED)
-            .trainer(trainer)
-            .ptContract(
-                PtContractEntity.builder()
-                    .id(1000L)
-                    .trainer(trainer)
-                    .trainee(trainee)
-                    .totalSession(10)
-                    .usedSession(5)
-                    .build()
-            )
-            .build())
-        );
-
-    //then
-    assertThrows(
-        ScheduleStartWithin1Day.class,
-        () -> scheduleService.cancelSchedule(
-            dto,
-            LocalDateTime.of(2024, 7, 15, 8, 0, 0) // 전날 아침 8시에 취소하려함
-        ));
   }
 }
